Test: move playwright loading page to shared helper, add axe tests
